Delete dockerfile.nanorc
Issue #47 brings up that there was already a `Dockerfile.nanorc` in the repo.  It being sorted in case-sensitive alphabetical order means that `Dockerfile.nanorc` was at the top of the listing, while `docerfile.nanorc` was sorted with the rest of the files.  Removing this as it is a duplicate (has the same contents) and people are having issues on non-case sensitive environments.